Add keyboard firmware
I liked to use all special keys from the keyboard, but the original
HID Library does not support a bunch of Special Keys. I found the
HID-Project and implemented the missing keys. But there is one big
issue, the sun_keyboard_map has to be rewritten to support the special
key codes, But I was to lazy to look at this issue, at the end I
implemented the Function directly and use the header only for the
not hard coded keys. It could be better, but for the first stage it
works and I currently wirting the commit message on the sun type 5
Keyboard.

This Version supports some of the functions on the left side, but
only on mac os x. A Windows Version is not planed.